requirements.txt
flask==3.0.3
groq==0.8.0
python-dotenv==1.0.1
